Tired of playing around with this....
This is the final commit.  This idea is dead because Python will hold
onto the file handle until it feels like garbage collecting it.  This
means when you try to spawn the subprocess, inkscape can't open the file
until the garbage collection occurs or the python.exe is totally dead.